Rearrange code to make it more manageable
Starting in a single file was fine, but maintainence will be a nightmare
if every single edit affects the same file. Instead, break things out
into multiple files grouping things in a meaningful way where it makes
sense.

Simultaneously, we keep index.ts and utilize it as a meta-export file
(aka "barrel" in Angular parlance) to make it easier for consumers to
import types related to the core functionality.
